let `rustc_smir` host `stable_mir` for refactoring
